Create README.md
English: This project conducted a high oxygen experiment on kuruma shrimp (Penaeus japonicus), with LS group as the control group and HS group as the high oxygen group. 
This project is RNA seq, and the sequencing sample is gills. The raw data was uploaded to NCBI as SRA data with the same name. This project is the process code for all 
data processing and is only for reproduction or learning purposes. It cannot be used for commercial purposes. Relevant articles have already been submitted.